refactor: Notice 컴포넌트를 GuideBox로 이름 변경
