+++ src/components/common/_ui/GuideBox/styles.module.scss
@@ -1,31 +1,47 @@
 @use '@/styles/variables' as *;
 
-.notice {
+.guideBox {
   background: $white;
-  padding: 15px 15px;
+  padding: 10px 15px;
   margin: 25px 0;
-  border-radius: 30px;
+  border-radius: 20px;
   border: 7px solid $blue-6;
   color: #333;
   text-align: center;
 
   .title {
     display: flex;
-    justify-content: space-between;
     align-items: center;
+    justify-content: center;
   }
 
-  h3 {
-    flex-grow: 1;
-    font-size: 18px;
-    font-weight: 700;
-    text-align: center;
-    color: #333;
-    margin-right: -10px;
+  .textWrapper {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex: 1;
+    gap: 4px;
+
+    svg {
+      color: #6b7280;
+    }
+
+    h3 {
+      font-size: 16px;
+      font-weight: 700;
+      text-align: center;
+      color: #444;
+      margin-right: -10px;
+    }
+  }
+
+  .arrowIcon {
+    margin-left: auto;
   }
 
   .content {
     margin-top: 5px;
+    font-size: 15px;
 
     p {
       color: #333;
