+++ src/components/common/index.ts
@@ -11,7 +11,7 @@ export { ErrorMessage } from './_ui/ErrorMessage/ErrorMessage'
 export { Logo } from './_ui/Logo/Logo'
 export { Button } from './_ui/Button/Button'
 export { Modal } from './_ui/Modal/Modal'
-export { Notice } from './_ui/Notice/Notice'
+export { GuideBox } from './_ui/GuideBox/GuideBox'
 
 export { Navbar } from './Navbar/Navbar'
 export { ProfileCard } from './ProfileCard/ProfileCard'
